Updated font, button, and made multi column
I converted the font format and then it worked. (I didn't know ttc didn't work until now.)
Removed the border radius on the button to blend in with the other links in the nav bar.
In the media query I added display flex for the items to be next to each other on larger screens.